+++ sregistry/client/pull.py
@@ -26,7 +26,6 @@ import os
 
 
 def main(args,parser,subparser):
-
     from sregistry.main import get_client, Client as cli
     
     # Does the user want to save the image?
@@ -46,7 +45,7 @@ def main(args,parser,subparser):
     for image in images:
 
         # Customize client based on uri
-        cli = get_client(image)
+        cli = get_client(image, quiet=args.quiet)
         cli.announce(args.command)
         response = cli.pull(images=image,
                             file_name=name,