+++ sregistry/main/base/__init__.py
@@ -34,7 +34,11 @@ from sregistry.main.base.headers import (
 
 from sregistry.main.base.http import ( 
     call, delete, download, get, paginate_get, 
-    post, put, stream, verify
+    post, put, stream, stream_response, verify
+)
+
+from sregistry.main.base.inspect import (
+    get_metadata
 )
 
 from sregistry.main.base.settings import (
@@ -66,21 +70,18 @@ class ApiConnection(object):
 # Metadata
 
     def speak(self):
-        '''a function for the client to announce him or herself, depending
-           on the level specified.
         '''
-        bot.info('[client|%s] [database|%s]' %(self.client_name,
-                                               self.database))
+           a function for the client to announce him or herself, depending
+           on the level specified. If you want your client to have additional
+           announced things here, then implement the class `_speak` for your
+           client.
 
-        self._speak()
-
-    def announce(self, command=None):
-        '''the client will announce itself given that a command is not in a
-           particular predefined list.
         '''
-        if command is not None:
-            if command not in ['get']:
-                self.speak()
+        if self.quiet is False:
+            bot.info('[client|%s] [database|%s]' %(self.client_name,
+                                                   self.database))
+
+            self._speak()
 
 
     def _speak(self):
@@ -91,6 +92,15 @@ class ApiConnection(object):
            download url (and nothing else)
         '''
         pass
+
+    def announce(self, command=None):
+        '''the client will announce itself given that a command is not in a
+           particular predefined list.
+        '''
+        if command is not None:
+            if command not in ['get'] and self.quiet is False:
+                self.speak()
+
 
     def __repr__(self):
         return "[client][%s]" %self.client_name
@@ -112,6 +122,9 @@ ApiConnection._get_setting = get_setting
 ApiConnection._get_and_update_setting = get_and_update_setting
 ApiConnection._get_storage_name = get_storage_name
 
+# Metadata
+ApiConnection.get_metadata = get_metadata
+
 # Auth
 ApiConnection.require_secrets = require_secrets
 ApiConnection._auth_flow = auth_flow
@@ -124,5 +137,6 @@ ApiConnection._get = get
 ApiConnection._paginate_get = paginate_get
 ApiConnection._post = post
 ApiConnection._put = put
-ApiConnection._stream = stream
+ApiConnection.stream = stream
+ApiConnection._stream = stream_response
 ApiConnection._verify = verify